holmesGabrielFinal
Main Menu Scene to begin with.

There were some things I was able to fix that were honestly major bugs and issues. Sadly I missed some of the art but id rather have it in on time than late

There is also a purple blueish orb behind a platform on level 2 it will end the game and you will complete your penance